% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures.R
\name{braycurtis}
\alias{braycurtis}
\title{Bray-Curtis dissimilarity}
\usage{
braycurtis(x, y)
}
\arguments{
\item{x, }{y}
}
\description{
Calculates the Bray-Curtis dissimilarity of two vectors x and y.
}
\examples{
x <- c()
y <- c()
braycurtis(x, y)
}
\seealso{
\code{\link[=canberra]{canberra()}}
}
