% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores.R
\name{brier}
\alias{brier}
\title{Brier score}
\usage{
brier(rev = FALSE, ...)
}
\arguments{
\item{rev}{whether to return 1 - Brier (TRUE) or not, defaults to FALSE}
}
\description{
The Brier score originates from weather forecasting. It is given by
\deqn{Brier = 1 / N \sum_{i = 1}^N (p_i - o_i)^2}
and is the mean squared error between the predicted probability, \eqn{p},
and the actual occurance, \eqn{o}.
}
\references{
\insertRef{Brier1971}{zostmod}
}
\seealso{
Other classification scores: \code{\link{auc}},
  \code{\link{bcr}}, \code{\link{dor}}, \code{\link{dp}},
  \code{\link{err}}, \code{\link{et}}, \code{\link{f1}},
  \code{\link{fai}}, \code{\link{fm}}, \code{\link{gain}},
  \code{\link{gl}}, \code{\link{gm}}, \code{\link{ignr}},
  \code{\link{jacc}}, \code{\link{kappa}},
  \code{\link{lr}}, \code{\link{mcc}}, \code{\link{op}},
  \code{\link{rand}}, \code{\link{rt}}, \code{\link{rus}},
  \code{\link{sm}}, \code{\link{ss}}, \code{\link{tss}}
}
\concept{classification scores}
