% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures.R
\name{canberra}
\alias{canberra}
\title{Canberra distance}
\usage{
canberra(x, y)
}
\arguments{
\item{x, }{y}
}
\description{
Calculates the Canberra distance of two vectors x and y.
}
\examples{
x <- c(1, 2, -3)
y <- c(6, 4, 5)
canberra(x, y)
}
\seealso{
\code{\link[=braycurtis]{braycurtis()}}
}
