% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures.R
\name{fai}
\alias{fai}
\title{Faith measure}
\usage{
fai(tp, tn, fp, fn, ...)
}
\arguments{
\item{tp}{True positives}

\item{tn}{True negatives}

\item{fp}{False positives}

\item{fn}{False negatives}
}
\description{
The Faith measure is given by
\deqn{(TP - FP - FN) / (TP + TN + FP + FN)}
where \eqn{TP} denotes true positives, \eqn{TN} denotes true negatives,
\eqn{FP} denotes false positives, \eqn{FN} denotes false negatives.
}
\examples{
fai(tp = 45, fp = 15, fn = 25, tn = 15)
}
\references{
\insertRef{Faith1983}{zostmod}
}
\seealso{
\code{\link[=rus]{rus()}}

Other classification scores: \code{\link{auc}},
  \code{\link{bcr}}, \code{\link{brier}},
  \code{\link{dor}}, \code{\link{dp}}, \code{\link{err}},
  \code{\link{et}}, \code{\link{f1}}, \code{\link{fm}},
  \code{\link{gain}}, \code{\link{gl}}, \code{\link{gm}},
  \code{\link{ignr}}, \code{\link{jacc}},
  \code{\link{kappa}}, \code{\link{lr}}, \code{\link{mcc}},
  \code{\link{op}}, \code{\link{rand}}, \code{\link{rt}},
  \code{\link{rus}}, \code{\link{sm}}, \code{\link{ss}},
  \code{\link{tss}}
}
\concept{classification scores}
\keyword{classif}
