% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores.R
\name{gl}
\alias{gl}
\title{Gower and Legendre}
\usage{
gl(tp, tn, fp, fn, theta = 1, ...)
}
\arguments{
\item{tp}{True positives}

\item{tn}{True negatives}

\item{fp}{False positives}

\item{fn}{False negatives}

\item{theta}{The weighting parameter, defaults to 1}
}
\description{
The Gower and Legandre coefficient is given by
\deqn{GL = (TP + TN) / (TP + TN + \theta (FP + FN))}
where \eqn{TP} denotes true positives, \eqn{TN} denotes true negatives,
\eqn{FP} denotes false positives, \eqn{FN} denotes false negatives, and
\eqn{\theta} is a weighting parameter. In our implementation \eqn{\theta}
has a default value of \eqn{1}, making it simple matching.
}
\examples{
g1(tp = 45, fp = 15, fn = 25, tn = 15)
}
\references{
\insertRef{GowerLegendre1986}{zostmod}
}
\seealso{
\code{\link[=sm() ss() rt]{sm() ss() rt()}}

Other classification scores: \code{\link{auc}},
  \code{\link{bcr}}, \code{\link{brier}},
  \code{\link{dor}}, \code{\link{dp}}, \code{\link{err}},
  \code{\link{et}}, \code{\link{f1}}, \code{\link{fai}},
  \code{\link{fm}}, \code{\link{gain}}, \code{\link{gm}},
  \code{\link{ignr}}, \code{\link{jacc}},
  \code{\link{kappa}}, \code{\link{lr}}, \code{\link{mcc}},
  \code{\link{op}}, \code{\link{rand}}, \code{\link{rt}},
  \code{\link{rus}}, \code{\link{sm}}, \code{\link{ss}},
  \code{\link{tss}}
}
\concept{classification scores}
\keyword{classif}
