% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores.R
\name{gm}
\alias{gm}
\title{Geometric mean}
\usage{
gm(tp, tn, fp, fn, adj = FALSE, ...)
}
\arguments{
\item{tp}{True positives}

\item{tn}{True negatives}

\item{fp}{False positives}

\item{fn}{False negatives}

\item{adj}{If TRUE calculate the adjusted mean, defaults to FALSE}
}
\description{
The geometric mean is given by
\deqn{GM = \sqrt(TP / (TP + FN) \cdot TN / (TN + FP))}
where \eqn{TP} denotes true positives, \eqn{TN} denotes true negatives,
\eqn{FP} denotes false positives, and \eqn{FN} denotes false negatives.
}
\details{
The adjusted geometric mean is given by
\deqn{AGM = (GM + (TN / (TN + FP)) + (FP + TN)) / (1 + FP + TN) if TP / (TP + FN) > 0}
and \eqn{0} otherwise, where \eqn{TP} again denotes true positives, \eqn{TN}
true negatives, \eqn{FP} false positives, and \eqn{FN} false negatives.
}
\examples{
gm(tp = 45, fp = 15, fn = 25, tn = 15)
}
\seealso{
Other classification scores: \code{\link{auc}},
  \code{\link{bcr}}, \code{\link{brier}},
  \code{\link{dor}}, \code{\link{dp}}, \code{\link{err}},
  \code{\link{et}}, \code{\link{f1}}, \code{\link{fai}},
  \code{\link{fm}}, \code{\link{gain}}, \code{\link{gl}},
  \code{\link{ignr}}, \code{\link{jacc}},
  \code{\link{kappa}}, \code{\link{lr}}, \code{\link{mcc}},
  \code{\link{op}}, \code{\link{rand}}, \code{\link{rt}},
  \code{\link{rus}}, \code{\link{sm}}, \code{\link{ss}},
  \code{\link{tss}}
}
\concept{classification scores}
\keyword{classif}
