% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores.R
\name{ignr}
\alias{ignr}
\title{Ignorance score}
\usage{
ignr(...)
}
\description{
The ignorance score, also referred to as the logarithmic score
}
\references{
\insertRef{Good1952}{zostmod}
\insertRef{RoulstonSmith2002}{zostmod}
}
\seealso{
Other classification scores: \code{\link{auc}},
  \code{\link{bcr}}, \code{\link{brier}},
  \code{\link{dor}}, \code{\link{dp}}, \code{\link{err}},
  \code{\link{et}}, \code{\link{f1}}, \code{\link{fai}},
  \code{\link{fm}}, \code{\link{gain}}, \code{\link{gl}},
  \code{\link{gm}}, \code{\link{jacc}},
  \code{\link{kappa}}, \code{\link{lr}}, \code{\link{mcc}},
  \code{\link{op}}, \code{\link{rand}}, \code{\link{rt}},
  \code{\link{rus}}, \code{\link{sm}}, \code{\link{ss}},
  \code{\link{tss}}
}
\concept{classification scores}
\keyword{classif}
