% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores.R
\name{kappa}
\alias{kappa}
\title{Cohen's kappa}
\usage{
kappa(tp, tn, fp, fn, ...)
}
\arguments{
\item{tp}{True positives}

\item{tn}{True negatives}

\item{fp}{False positives}

\item{fn}{False negatives}
}
\description{
Cohen's kappa (\eqn{\kappa}), also known as Heidke's skill score, is given by
\deqn{\kappa = 1 - (1 - p_o) / (1 - p_e)}
where \eqn{p_o} is the observed frequency and \eqn{p_e} is the
expected frequency. We calculate \eqn{p_o} and \eqn{p_e} based on true positives
(\eqn{TP}), true negatives (\eqn{NP}), false positives (\eqn{FP}), and
false negatives (\eqn{FN}).
}
\examples{
kappa(tp = 45, fp = 15, fn = 25, tn = 15)
}
\references{
\insertRef{Cohen1960}{zostmod}
\insertRef{Heidke1926}{zostmod}
}
\seealso{
\code{\link[=tss]{tss()}}

Other classification scores: \code{\link{auc}},
  \code{\link{bcr}}, \code{\link{brier}},
  \code{\link{dor}}, \code{\link{dp}}, \code{\link{err}},
  \code{\link{et}}, \code{\link{f1}}, \code{\link{fai}},
  \code{\link{fm}}, \code{\link{gain}}, \code{\link{gl}},
  \code{\link{gm}}, \code{\link{ignr}}, \code{\link{jacc}},
  \code{\link{lr}}, \code{\link{mcc}}, \code{\link{op}},
  \code{\link{rand}}, \code{\link{rt}}, \code{\link{rus}},
  \code{\link{sm}}, \code{\link{ss}}, \code{\link{tss}}
}
\concept{classification scores}
\keyword{classif}
