% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores.R
\name{mcc}
\alias{mcc}
\title{Matthew's correlation coefficient}
\usage{
mcc(tp, tn, fp, fn, ...)
}
\arguments{
\item{tp}{True positives}

\item{tn}{True negatives}

\item{fp}{False positives}

\item{fn}{False negatives}
}
\description{
Matthew's correlation coefficient (\eqn{MCC}) is given by
\deqn{MCC = (TP \cdot TN - FP \cdot FN) / (\sqrt((TP + FP)(TP + FN)(TN + FP)(TN + FN)))}
where \eqn{TP} denotes true positives, \eqn{TN} denotes true negatives,
\eqn{FP} denotes false positives, and \eqn{FN} denotes false negatives.
}
\details{
It is considered to be better than F1 and Rand index (accuracy) for binary classification
\insertCite{@see @Chicco2017 for details}{zostmod}.
}
\examples{
mcc(tp = 45, fp = 15, fn = 25, tn = 15)
}
\references{
\insertRef{Chicco2017}{zostmod}
}
\seealso{
\code{\link[=f1() rand]{f1() rand()}}

Other classification scores: \code{\link{auc}},
  \code{\link{bcr}}, \code{\link{brier}},
  \code{\link{dor}}, \code{\link{dp}}, \code{\link{err}},
  \code{\link{et}}, \code{\link{f1}}, \code{\link{fai}},
  \code{\link{fm}}, \code{\link{gain}}, \code{\link{gl}},
  \code{\link{gm}}, \code{\link{ignr}}, \code{\link{jacc}},
  \code{\link{kappa}}, \code{\link{lr}}, \code{\link{op}},
  \code{\link{rand}}, \code{\link{rt}}, \code{\link{rus}},
  \code{\link{sm}}, \code{\link{ss}}, \code{\link{tss}}
}
\concept{classification scores}
\keyword{classif}
