% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures.R
\name{rt}
\alias{rt}
\title{Rogers-Tanimoto}
\usage{
rt(tp, tn, fp, fn, ...)
}
\arguments{
\item{tp}{True positives}

\item{tn}{True negatives}

\item{fp}{False positives}

\item{fn}{False negatives}
}
\description{
The Rogers-Tanimoto coefficient is given by
\deqn{RT = (TP + TN) / (TP + TN + 2FP + 2FN)}
where \eqn{TP} denotes true positives, \eqn{TN} denotes true negatives,
\eqn{FP} denotes false positives, \eqn{FN} denotes false negatives.
}
\examples{
rt(tp = 45, fp = 15, fn = 25, tn = 15)
}
\seealso{
\code{\link[=gl() sm() ss]{gl() sm() ss()}}

Other classification scores: \code{\link{auc}},
  \code{\link{bcr}}, \code{\link{brier}},
  \code{\link{dor}}, \code{\link{dp}}, \code{\link{err}},
  \code{\link{et}}, \code{\link{f1}}, \code{\link{fai}},
  \code{\link{fm}}, \code{\link{gain}}, \code{\link{gl}},
  \code{\link{gm}}, \code{\link{ignr}}, \code{\link{jacc}},
  \code{\link{kappa}}, \code{\link{lr}}, \code{\link{mcc}},
  \code{\link{op}}, \code{\link{rand}}, \code{\link{rus}},
  \code{\link{sm}}, \code{\link{ss}}, \code{\link{tss}}
}
\concept{classification scores}
\keyword{classif}
