% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures.R
\name{tss}
\alias{tss}
\title{True skill statistic}
\usage{
tss(tp, tn, fp, fn, ...)
}
\arguments{
\item{tp}{True positives}

\item{tn}{True negatives}

\item{fp}{False positives}

\item{fn}{False negatives}
}
\description{
The true skill statistic (\eqn{TSS}) also known as Youden's index or markedness
is a measure often used in ecology for assessing species distribution models.
It is given by
\deqn{TSS = (TP / (TP + FN)) + (TN / (TN + FP)) - 1}
where \eqn{TP} denotes true positives, \eqn{TN} denotes true negatives,
\eqn{FP} denotes false positives, and \eqn{FN} denotes false negatives.
}
\examples{
tss(tp = 45, fp = 15, fn = 25, tn = 15)
}
\references{
\insertRef{Youden1950}{zostmod}
}
\seealso{
\code{\link[=kappa]{kappa()}}

Other classification scores: \code{\link{auc}},
  \code{\link{bcr}}, \code{\link{brier}},
  \code{\link{dor}}, \code{\link{dp}}, \code{\link{err}},
  \code{\link{et}}, \code{\link{f1}}, \code{\link{fai}},
  \code{\link{fm}}, \code{\link{gain}}, \code{\link{gl}},
  \code{\link{gm}}, \code{\link{ignr}}, \code{\link{jacc}},
  \code{\link{kappa}}, \code{\link{lr}}, \code{\link{mcc}},
  \code{\link{op}}, \code{\link{rand}}, \code{\link{rt}},
  \code{\link{rus}}, \code{\link{sm}}, \code{\link{ss}}
}
\concept{classification scores}
\keyword{classif}
